8-23
인터페이스 작동 수정.

몬스터 생성시 기본 prepab을 받은뒤 스테이터스 table을 받아서 작동하도록 변경

+++ Assets/Script/Manager/SaveLoadManager.cs
@@ -61,7 +61,7 @@ public class GameManagerSaveData
 {
     public int stage;
     public int round;
-    public int lampLight;
+    public float lampLight;
     public List<string> roundInfo = new List<string>();
